Make read/write device state handled by touch_sensor
