+++ src/utils/utils.py
@@ -3,6 +3,9 @@ import fcntl
 import subprocess
 import logging
 import os
+import json
+import time
+from typing import Dict, Any
 
 from src.utils.paths import PROJECT_DIR
 
@@ -52,3 +55,71 @@ def get_device_id() -> str:
         return process.stdout.strip()
     except subprocess.CalledProcessError as e:
         logger.error(f"An error occurred while fetching the hostname: {e}")
+
+
+def is_at_least_next_day(from_timestamp: float, to_timestamp: float) -> bool:
+    """Check if the 'to_timestamp' is at least the next day compared to 'from_timestamp'.
+
+    Args:
+        from_timestamp (float): The starting Unix timestamp.
+        to_timestamp (float): The ending Unix timestamp.
+
+    Returns:
+        bool: True if 'to_timestamp' is at least the next day from 'from_timestamp', False otherwise.
+    """
+    day_in_seconds = 86400
+    # Get the start of the day for both timestamps
+    from_day_start = from_timestamp - (from_timestamp % day_in_seconds)
+    to_day_start = to_timestamp - (to_timestamp % day_in_seconds)
+
+    # Check if to_timestamp is at least the next day from from_timestamp
+    return to_day_start > from_day_start
+
+
+def read_device_state(path_to_device_state: str) -> Dict[str, Any]:
+    """Read the device state from a JSON file.
+
+    Args:
+        path_to_device_state (str): Path to the JSON file containing the device state.
+
+    Returns:
+        Dict[str, Any]: A dictionary representing the device state.
+    
+    Raises:
+        FileNotFoundError: If the specified JSON file path does not exist.
+        JSONDecodeError: If there is an error decoding the JSON data from the file.
+    """
+    logger.info('Reading device state')
+    try:
+        with open(path_to_device_state, 'r') as file:
+            return json.load(file)
+    except FileNotFoundError:
+        logger.error('Device state file not found')
+        raise FileNotFoundError
+    except json.JSONDecodeError:
+        logger.error('Unable to parse device state: Invalid JSON format')
+        raise json.JSONDecodeError
+
+
+def write_device_state(device_state: Dict[str, Any], path_to_device_state: str) -> None:
+    """Write the updated device state to a JSON file.
+
+    Args:
+        device_state (Dict[str, Any]): The device state to write.
+        path_to_device_state (str): Path to the JSON file where the device state will be saved.
+    
+    Raises:
+        FileNotFoundError: If the specified JSON file path does not exist.
+        IOError: If there is an error writing to the file.
+    """
+    logger.info('Writing device state')
+    device_state['saved_timestamp'] = time.time()
+    try:
+        with open(path_to_device_state, 'w') as file:
+            json.dump(device_state, file, indent=4)
+    except FileNotFoundError:
+        logger.error('Device state file not found')
+        raise FileNotFoundError  # TODO: Determine error message format
+    except IOError as e:
+        logger.error(f"An error occurred while writing to device state: {e}")
+        raise IOError  # TODO: Determine error message format
